src/application/use-cases/educational-content-cases/trail/read/read-published/read-published-trails-use-case.ts
import {
    ITrailRepository,
    ReadPublishedTrailsUseCaseOutputDTO,
    Trail
} from "./index";


export class ReadPublishedTrailsUseCase {
    constructor(private trailRepository: ITrailRepository) { }

    async execute(): Promise<ReadPublishedTrailsUseCaseOutputDTO> {
        const publishedTrails: Trail[] = await this.trailRepository.listPublished();

        const outputDTO: ReadPublishedTrailsUseCaseOutputDTO = {
            trails: publishedTrails
        }

        return outputDTO;
    }
}       
